+++ tests/UnitTests/TestInprocessNode.cpp
@@ -267,7 +267,7 @@ TEST_F(InProcessNodeTests, getBlocksByHeightEmpty) {
 }
 
 TEST_F(InProcessNodeTests, getBlocksByHeightMany) {
-  const uint32_t NUMBER_OF_BLOCKS = 10;
+  const size_t NUMBER_OF_BLOCKS = 10;
 
   std::vector<uint32_t> blockHeights;
   std::vector<std::vector<CryptoNote::BlockDetails>> actualBlocks;
@@ -311,7 +311,7 @@ TEST_F(InProcessNodeTests, getBlocksByHeightMany) {
 }
 
 TEST_F(InProcessNodeTests, getBlocksByHeightFail) {
-  const uint32_t NUMBER_OF_BLOCKS = 10;
+  const size_t NUMBER_OF_BLOCKS = 10;
 
   std::vector<uint32_t> blockHeights;
   std::vector<std::vector<CryptoNote::BlockDetails>> actualBlocks;
@@ -366,7 +366,7 @@ TEST_F(InProcessNodeTests, getBlocksByHashEmpty) {
 }
 
 TEST_F(InProcessNodeTests, getBlocksByHashMany) {
-  const uint32_t NUMBER_OF_BLOCKS = 10;
+  const size_t NUMBER_OF_BLOCKS = 10;
 
   std::vector<Crypto::Hash> blockHashes;
   std::vector<CryptoNote::BlockDetails> actualBlocks;
@@ -407,7 +407,7 @@ TEST_F(InProcessNodeTests, getBlocksByHashMany) {
 }
 
 TEST_F(InProcessNodeTests, getBlocksByHashFail) {
-  const uint32_t NUMBER_OF_BLOCKS = 10;
+  const size_t NUMBER_OF_BLOCKS = 10;
 
   std::vector<Crypto::Hash> blockHashes;
   std::vector<CryptoNote::BlockDetails> actualBlocks;
